requirements.txt
matplotlib~=3.5.1
setuptools~=61.2.0
numpy~=1.22.3
tqdm~=4.64.0
pyyaml~=6.0